+++ modules/Food_Service/includes/DeleteTransactionItem.fnc.php
@@ -1,9 +1,15 @@
 <?php
 
 /**
- * @param $transaction_id
- * @param $item_id
- * @param $type
+ * Delete Transacion Item
+ *
+ * @since 10.0 Use db_trans_*() functions
+ *
+ * @uses DeleteTransaction()
+ *
+ * @param $transaction_id Transaction ID.
+ * @param $item_id        Item ID.
+ * @param $type           'student' or 'staff'.
  */
 function DeleteTransactionItem( $transaction_id, $item_id, $type = 'student' )
 {
@@ -20,7 +26,11 @@ function DeleteTransactionItem( $transaction_id, $item_id, $type = 'student' )
 		$sql3 = "DELETE FROM FOOD_SERVICE_TRANSACTION_ITEMS WHERE TRANSACTION_ID='" . (int) $transaction_id . "' AND ITEM_ID='" . (int) $item_id . "'";
 	}
 
-	DBQuery( 'BEGIN; ' . $sql1 . '; ' . $sql2 . '; ' . $sql3 . '; COMMIT' );
+	db_trans_start();
+
+	db_trans_query( $sql1 . '; ' . $sql2 . '; ' . $sql3 . ';' );
+
+	db_trans_commit();
 
 	//FJ if no more transaction items, delete transaction
 	$trans_items_RET = DBGet( "SELECT ITEM_ID FROM " . ( $type == 'staff' ? "food_service_staff_transaction_items" : "FOOD_SERVICE_TRANSACTION_ITEMS" ) . " WHERE TRANSACTION_ID='" . (int) $transaction_id . "'" );